Remove ExtensionContextCreator
