--- src/Schema/Extension/ExtensionContextCreatorInterface.php
@@ -1,16 +1,0 @@
-<?php
-
-namespace Digia\GraphQL\Schema\Extension;
-
-use Digia\GraphQL\Language\Node\DocumentNode;
-use Digia\GraphQL\Schema\SchemaInterface;
-
-interface ExtensionContextCreatorInterface
-{
-    /**
-     * @param SchemaInterface $schema
-     * @param DocumentNode    $document
-     * @return ExtensionContextInterface
-     */
-    public function create(SchemaInterface $schema, DocumentNode $document): ExtensionContextInterface;
-}
