+++ src/Schema/Extension/SchemaExtensionProvider.php
@@ -11,7 +11,6 @@ class SchemaExtensionProvider extends AbstractServiceProvider
      * @var array
      */
     protected $provides = [
-        ExtensionContextCreatorInterface::class,
         SchemaExtenderInterface::class,
     ];
 
@@ -20,10 +19,7 @@ class SchemaExtensionProvider extends AbstractServiceProvider
      */
     public function register()
     {
-        $this->container->add(ExtensionContextCreatorInterface::class, ExtensionContextCreator::class, true)
-            ->withArgument(DefinitionBuilderCreatorInterface::class);
-
         $this->container->add(SchemaExtenderInterface::class, SchemaExtender::class)
-            ->withArgument(ExtensionContextCreatorInterface::class);
+            ->withArgument(DefinitionBuilderCreatorInterface::class);
     }
 }
